refactor(pty): #983 spawn の telemetry ヘルパを spawn_metrics.rs へ分離
spawn.rs が file-size baseline (785行) を 797行で超過し、File-size ratchet が
main 上で失敗していた。spawn 本体の制御フローから独立した純粋関数群
(build_cmd_label / engine_label / platform_label / log_spawn_outcome) を
新規 spawn_metrics.rs へ移設し、spawn.rs を 730 行に縮小する。

挙動・ログフォーマットは一切変更なし (純粋な move)。

Closes #983
